requirements.txt
moviepy==1.0.3
pydub==0.25.1
python-dotenv==1.0.1
srt==3.5.3
streamlit==1.38.0
vosk==0.3.45
